Avoid posting empty receipts by making`TransactionsManager` always use `SK1` implementation (#2015)
Fixes [CSDK-513].

The SDK still uses `SK1` when it comes to fetching receipts to post to
the backend.
The `SK1` implementation checks whether there are transactions inside of
the receipt.

However, this could lead to issues if `SK2` has transactions but SK1
doesn't. In those cases, we would end up posting an empty receipt. This
PR fixes that.

[CSDK-513]:
https://revenuecats.atlassian.net/browse/CSDK-513?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ