`ErrorCode.unknownBackendError`: include original error code
Currently we only include the message and not the error code. So when these unknown errors are returned, unless the message is clear, we might not known which code the backend actually sent.
Example:
> There was an unknown backend error. The purchased product was missing in the receipt. This is typically due to a bug in StoreKit.

In this case it's clear, but nowhere in the logs we can find the error code.

This adds the error code to those messages (only when the code is unknown), as well as to `userInfo`.

### Examples from the tests

#### Before:
> There was an unknown backend error. This is a future unknown error
#### After:
> There was an unknown backend error. This is a future unknown error (1234)

#### Before:
> 😿‼️ There was an unknown backend error. Page not found
#### After:
> 😿‼️ There was an unknown backend error. Page not found (6000)