`CachingProductsManager`: use partial cached products (#2014)
#1907 introduced this `CachingProductsManager` abstraction, but simply
copied the existing caching mechanism, which was very limited. This
expands on it by not ignoring cached products when not _all_ requested
products are found. Instead, it uses the partial cache hit, and requests
the rest of products.

This performance optimization is illustrated in the test change, as the
second request only fetches the missing product now.