`Integration Tests`: changed weekly to monthly subscriptions to work around 0-second subscriptions
As exposed by #2038, weekly subscriptions were leading to false negatives because they were always expired by the time the purchase was made using `TimeRate.monthlyRenewalEveryThirtySeconds`, instead of lasting 10 seconds as documented.
(`FB11767559`)

The advantage of using weekly subscriptions was to also work around `FB11767679`, which means that we have to manually _wait_ for the subscription to expire, as `SKTestSession.expireSubscription(productIdentifier:)` doesn't actually change the receipt.
Because we had to wait, using weekly subscriptions was faster, since we only needed to wait 10 seconds.

This does mean that expirations in these tests now have to wait 30 seconds instead of 10, but now they're no longer flaky :tada: