`HTTPClient`: added log for failed requests
I'm looking into [CSDK-517], and it's hard to tell why we ended up with concurrent requests. It appears that a second request begins before the last one started.
That's seemingly impossible, and the only explanation for the logs missing "request completed" is that the first request failed.

This extra log will help us understand what's happening.