`Tests`: replaced `toEventually` with new `waitUntilValue` to simplify tests
Example:
```diff
-        var result: Result<Offerings, OfferingsManager.Error>?
-        offeringsManager.offerings(appUserID: MockData.anyAppUserID) {
-            result = $0
+        let result = waitUntilValue { completed in
+            self.offeringsManager.offerings(appUserID: MockData.anyAppUserID) {
+                completed($0)
+            }
         }
-        expect(result).toEventuallyNot(beNil())
```

This removes the need to _poll_ for when the value is returned, and instead it gets detected immediately, potentially providing a significant boost in test execution speed.
I've also used `Nimble`'s own `waitUntil` for cases where we don't need to return a value.

 #### `toEventually` usage:
- Before this PR: 353
- After: 229

There's still a lot of legitimate usage of `toEventually`, but I'm sure I missed other places where this can also simplify the code.
This is a good start for now.