`Configuration.with(appUserID:)`: changed parameter to `String?`
The docs had this:
> Pass `nil` or an empty string if you want ``Purchases`` to generate this for you.

But `nil` wasn't a valid input.
I noticed [this reported issue](https://github.com/RevenueCat/purchases-ios/issues/2107#issuecomment-1332007011) where the user was doing this:
```swift
.with(appUserID: userUID ?? "")
```

Which technically would have the same semantics, since we allow empty and convert it to anonymous, but now this is more clear and it matches the docstring.