`Integration Tests`: fixed `Purchases` leak through `PurchasesDiagnostics`
Calling `PurchasesDiagnostics.default` was lazily initializing it:
```swift
public static let `default`: PurchasesDiagnostics = .init(purchases: Purchases.shared)
```

Which was keeping a reference to `Purchases.shared` around.
For most cases this is fine, since `Purchases` would only be initialized once, but for tests it meant that it was leaking, which is now detected by #2106.