Fixed race condition finishing `SK1` transactions
Fixes [CSDK-504].

`SKPaymentQueue.finishTransaction` is asynchronous, but we were assuming it wasn't.
This meant that when performing multiple purchases (like consumables) in quick succession, before the first transaction was finished, the second purchase was reusing the first transaction.

Example log:
```
[Purchases] - INFO: :moneybag: Finishing transaction ‘0’ for product ‘consumable.10_coins’
[Purchases] - INFO: :heart_eyes_cat::moneybag: Purchased product - ‘consumable.10_coins’
[Purchases] - INFO: :moneybag: Purchasing Product ‘consumable.10_coins’ from package in Offering ‘coins’
[Purchases] - DEBUG: :information_source: StoreKit1Wrapper (0x00006000008a9530) removedTransaction: consumable.10_coins 0 1
```

Notice how the "removed transaction" was notified _after_ we had already started the second purchase.