Added integration and unit tests to verify observer mode behavior (#2069)
See [TRIAGE-179].

I wrote all these to cover the change in #2063. I wrote a failing test
that only passed with the change in #2063. However, when running that in
SK1, that would fail.
With the change as suggested in #2063, the SK2 behavior would differ
from SK1.

From the `Transaction.all` docs:
> This sequence returns the user’s transaction history current to the
moment you access the sequence. The sequence emits a finite number of
transactions. If the App Store processes new transactions for the user
while you’re accessing this sequence, the new transactions appear in the
transaction listener, updates.

These new tests verify the behavior of the SDK when encountering
transactions created prior to initializing the SDK, and some made after
initialization.

Integration tests now are as follows:
- `BaseStoreKitIntegrationTests`
  - `StoreKit1IntegrationTests`
  - `StoreKit2IntegrationTests`
  - `BaseStoreKitObserverModeIntegrationTests`:
    - `StoreKit1ObserverModeIntegrationTests`
    - `StoreKit2ObserverModeIntegrationTests`
    - `StoreKit1ObserverModeWithExistingPurchasesTests`
    - `StoreKit2ObserverModeWithExistingPurchasesTests`

Depends on #2064, #2066, and #2134.

[TRIAGE-179]:
https://revenuecats.atlassian.net/browse/TRIAGE-179?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ