`DNSChecker`: added log with resolved host
Example:
> DEBUG: ℹ️ Failing url 'https://0.0.0.0/offers' resolved to host '0.0.0.0'

Looking into #2028, it appears that the user is getting a `NSURLErrorCannotFindHost` with a `NSErrorFailingURLKey`.
We didn't determine that this was a `DNS` error because whatever the host resolved to, wasn't one of `Set(["0.0.0.0", "127.0.0.1"])`.

This will add a log that will help us figure out why they got a `NSURLErrorCannotFindHost`.