`Purchase Tester`: added ability to purchase products directly with `StoreKit` (#2172)
This will help test observer mode, mimicking how an app might use
`RevenueCat` with observer mode.

When that setting is enabled, 2 new buttons will appear:
![Simulator Screen Shot - iPhone 14 Pro Max - 2022-12-20 at 08 58
07](https://user-images.githubusercontent.com/685609/208723050-56644a9a-d78f-4788-bda1-b64bdc400b40.png)

The implementation isn't the most maintainable, but we don't really need
it to be. I favored simplicity instead of over-complicating the
implementation. It's not meant to be a canonical example of the correct
way to purchase with StoreKit, just a quick way of testing this
behavior.