`PostReceiptDataOperation`: replaced receipt base64 with hash for cache key
Should fix #2198 and [SDKONCALL-200]. This is a big performance optimization in time and memory, as we no longer need to convert the entire receipt to base 64 and concatenate it to the string, which is what was likely causing #2198.
Now, the implementation of `Data.uuid` avoids any data copies and efficiently creates a `UUID` hash with the receipt content.

Apart from the new tests for this property, `BackendPostReceiptDataTests.testDoesntCacheForDifferentReceipts` ensures that different receipts have different cache keys, and therefore this new implementation still avoids collisions (other than the extremely unlikely hash collisions).