`CustomerInfo`: support parsing schema version 2
Fies [TRIAGE-240].

When we converted `CustomerInfo` to use `Codable` (#1496), the schema version was also changed from 2 to 3.
I don't remember if this was done on purpose, but turns out that it was not necessary.

This means that a user opening an app while offline, after upgrading from version `3.x` to `4.x` (>= `4.3`), they would not have a cached `CustomerInfo` and therefore no entitlements.
This fixes that by restoring support and adding a specific test with a sample version 2 JSON (which is virtually identical).