`Purchases`: avoid potential crash when initializing in the background (#2231)
#2224 likely fixes #1877. This change also avoids running the code being
triggered in that crashing stack trace, since it's not necessary.
`Purchases.delegate` setter already calls
`sendCachedCustomerInfoToDelegateIfExists`, so that method was
unnecessary. Removing this call to `sendCachedCustomerInfoIfAvailable`
doesn't make any test fails because of that.

This behavior is also verified by `PurchasesConfiguringTests`:
`testFirstInitializationFromBackgroundCallsDelegateForAnonIfInfoCached`
and `testSettingTheDelegateAfterInitializationSendsCachedCustomerInfo`.

I also improved this by extracting it into its own method and adding an
explanation for the `isApplicationBackgrounded` check.