`Purchases`: fixed documentation warnings
Fixes this warning:

These started happening after #1958. I assumed it was an Xcode bug, because they were showing up on the wrong file.
Turns out that by itself was an Xcode bug (filed as `FB11959877`):

But the underlying warning was correct, which this fixes. We just needed to fix the reference to `logIn` to point to the specific overload.