`PurchasesOrchestrator`: changed `ReceiptRefreshPolicy.always` to `.onlyIfEmpty` after a purchase
We've had reports of receipt fetching throttling errors (#2116). #2146 improved this by throttling receipt refreshing.
However, we can avoid this altogether by not refreshing the receipt unless it's empty.

The main reason being that the backend will automatically refresh the receipt using `/verifyReceipt`, so it doesn't matter if the receipt we post is slightly out of date.