Replaced `Any` uses for workaround with `Box`
See https://github.com/apple/swift/issues/58099
This is fixed in newer versions of Xcode, but we still need the workaround for older ones. Instead of using `Any` and force-casting, `Box` is better since it's type safe.

I've also updated the comments to point to the new `Swift` GitHub issue page.