`OfferingsManager`: ensure underlying `OfferingsManager.Error.configurationError` is logged
Changes this:
> ERROR: 🍎‼️ Error fetching offerings - The operation couldn't be completed. (RevenueCat.OfferingsManager.Error error 1.)

Into this:
> 🍎‼️ Error fetching offerings - The operation couldn’t be completed. (RevenueCat.OfferingsManager.Error error 1.)
There are no products registered in the RevenueCat dashboard for your offerings. If you don't want to use the offerings system, you can safely ignore this message. To configure offerings and their products, follow the instructions in https://rev.cat/how-to-configure-offerings.
More information: https://rev.cat/why-are-offerings-empty

This was likely a regression at some point, leading to a confusing error message that was missing the underlying `message` inside of `OfferingsManager.Error.configurationError`.