`HTTPClientTests`: fixed failing test with missing assertions (#2262)
`HTTPClientTest.testRequestIsRetriedIfResponseFromETagManagerIsNil` was
actually failing (the request wasn't successful), but because it had no
expectations, we didn't notice until we introduced Code Coverage.
As you can see [in the Codecov
diff](https://app.codecov.io/gh/RevenueCat/purchases-ios/pull/2258):
<img width="939" alt="Screenshot 2023-02-06 at 09 11 37"
src="https://user-images.githubusercontent.com/685609/217038326-b94e7023-701f-451c-8347-d6777e86ee39.png">

Most of the time that code didn't have coverage (meaning we weren't
testing it). The reason for why is uncovered in this PR.
The random execution order meant that sometimes (most of the time) tests
like `testHandlesRealErrorConditions` were running first, which did
this:
```swift
let response = HTTPStubsResponse.emptySuccessResponse
response.error = error
return response
```

That seemingly innocuous code was modifying
`HTTPStubsResponse.emptySuccessResponse`, defined in an extension in
that file. Being used to value types nobody thought much of that, but
unfortunately `HTTPStubsResponse` is a class, therefore it's a mutable
type with reference semantics. Because of that, every test executed
_after_ any of these modifying the response would receive a failure, but
it was going unnoticed.

This PR also introduces a couple improvements to how we wait for several
concurrent requests (which I thought was the issue at first).