`PurchaseOrchestrator`: always refresh receipt purchasing in sandbox (#2280)
Fixes #2260 and [SDKONCALL-216].
Follow up to #2245.

The change in #2245 was meant to avoid receipt refresh throttling errors
(#2116).
However, that was a regression when using StoreKit config files because
those don't get refreshed in the backend.

Unfortunately the workaround introduced in #1945 prevented us from
noticing this, because we use
`DangerousSettings.InternalSettings.enableReceiptFetchRetry` in our own
integration tests.
I added an integration test that explicitly covers this scenario. That
fails without this change if disabling that dangerous setting. I thought
about making that test run without the setting, but that might lead to
flaky failures in CI, which is why we introduced that workaround in the
first place. But at least now we have an explicit integration test for
this, as well as 2 unit tests.


[SDKONCALL-216]:
https://revenuecats.atlassian.net/browse/SDKONCALL-216?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ