`TestLogHandler`: increased default capacity
This fixes flaky test failures.
`TestLogHandler` wires into the global log handler. When running test concurrently, depending on which tests and how many are running, there could be a lot of logs being stored in there.
Specifically, `StoreKit1WrapperTests.testUpdatedTransactionsLogsWarningWhenSendingTooManyTransactions` logs 100 messages to test that too many transactions lead to a warning. With the current limit, it only took 50 other logs to trigger this `precondition`, which could happen relatively easily.