Add lane for running iOS v3 load shedder integration tests (#2388)
This adds a fastlane lane to execute load shedder integration tests. 
In a follow-up PR I'll call this lane automatically from CircleCI, on a
schedule.

It won't be a part of regular integration tests, since this logic is
tied to v3 and static. So only the load shedder itself is getting tested
here.