`DispatchTimeInterval`: fixed Xcode 14.3 compilation
I expected it to be marked `Sendable` in the next version of Swift, but it still hasn't so I filed https://github.com/apple/swift/issues/65044.
`DispatchTimeInterval` is a value type, so it's safe to do this. I don't want to remove the `#if swift` version check because I expect it to be marked `Sendable` in a future version.