Fix deadlock when sending notification center in customerInfoListener (#2407)
Fixes an issue where a deadlock could happen if you fire a notification
in the customerInfoListener.

The issue stems from our usage of NotificationCenter to detect deletion
of userDefaults.

Since we've changed the default UserDefaults anyway, I don't feel like
this adds much value anymore, and we've spent wayyyy too much time
dealing with race conditions caused by it, so I'm just getting rid of
it.

I couldn't get it to reproduce on my device, but @codykerns was able to
get a consistent repro and confirmed the fix with this branch.