Custom Entitlement Computation: avoid `getCustomerInfo` requests in `PurchasesOrchestrator` for cancelled purchases
Fixes SDK-3088
Supersedes #2447

Note that it's currently impossible to test this for SK2, so we only have SK1 tests (we have a Radar for that).

The next step will be to deprecate the current API to remove the `userCancelled` property and make this the new behavior.
See also #1910 for an explanation of the different return values of the purchase APIs.