`DangerousSettings`: debug-only `forceServerErrors`
Follow up to #2370. This removes `forceServerErrors` from release builds so it can't be accidentally used (even though `HTTPClient` was only checking it on `DEBUG` already).
Also allows changing the flag at runtime, so future integration tests can check the behavior when the server changes status throughout the lifetime of the app.