`DeviceCache`: refactored cache keys
Just a small refactor to make user-dependent keys require it at the type level instead of having concatenation of the keys everywhere.
This still could use much more refactoring to make the code more maintainable, but this is just a small change to make it easier to add new keys.