`HTTPClient`: don't assume error responses are JSON
When receiving 500 responses, `HTTPClient` was always calling `ErrorResponse.from(_ data: Data)`, which tried to decode the content as if it was JSON.
However, if the response had something completely different, it would fail to decode.

This wasn't really a bug because `HTTPClient` was still forwarding the right result, but now it will do so without even trying to parse it.