`Integration Tests`: workaround for `XCTest` crash after a test failure
See https://app.circleci.com/jobs/github/RevenueCat/purchases-ios/69881

Assertion failures in that method were leading to _crashes_ in `Purchases.shared`, instead of a simple test failure that could be retried.

For some reason, `XCTest` is continuing execution after an assertion failure in an `async` method, despite having `self.continueAfterFailure = false`.
This might even be a Swift mis-compilation. This PR is a workaround: instead of relying on `XCTest` stopping execution of the test, we instead throw an `Error` to force that at compile time.

Note that this still produces nice errors when an assertion fails inside of `verifyEntitlementWentThrough`: