`BasePurchasesTests`: fixed leak detection (#2534)
While working on #2533, I knew that my proof of concept had a retain
cycle. I however was very confused when it wasn't detected by this
mechanism introduced first introduced in #2104.

Turns out that this had been wrong the whole time: the current
implementation, as explained by the comment, was executing those blocks
in LIFO order, which meant that the references were being set to `nil`
before the assertions could be created.
This meant that by the time the `except { [weak purchases =
self.purchases] ... }` lines were called, `self.purchases` itself was
`nil`, so nothing was being checked.

This simplifies the implementation by inlining the checks, and saving a
`weak` reference before clearing the singleton and all the local
references.

Luckily there was only one failing test that had to do with how
`Purchases` was being created, and implementation details of
`expectFatalError`. To avoid this, I changed the test to use the
`Purchases.init` method directly.