Extracted `TransactionPoster` from `PurchasesOrchestrator` (#2540)
Necessary refactor for #2533. This allows `CustomerInfoManager` to have
an instance of `TransactionPosterType` instead of the whole
`PurchasesOrchestrator`.

Essentially this becomes the hierarchy now:

![IMG_B8C56C11B7E5-1](https://github.com/RevenueCat/purchases-ios/assets/685609/36f5c30b-fc82-4590-b574-ce6d95eef13f)

### Changes:
- Created stateless `TransactionPoster`
- Extracted `handlePurchasedTransaction` into `TransactionPoster`
- The new implementation there is _mostly_ the same, with the one
difference that it's stateless. Instead of holding
`purchaseCompleteCallbacksByProductID`, that's done by
`PurchasesOrchestrator`. The methods there all take a callback, so
they're easier to use and don't require keeping a callback around.
- Created `PurchaseSource` to abstract the combo of `isRestore` and
`InitiationSource`. This is still messy using the deprecated
`allowSharingAppStoreAccount`, but at least that's abstracted out and
the new `TransactionPoster` won't deal with that deprecated property.
- I didn't mock `TransactionPoster`, which means that all existing tests
that check `PurchasesOrchestrator` continue working the same way. This
was on purpose to keep this refactor as simple as possible.
- This supersedes #2536: the change made there was now possible here
without any integration test failures, thanks to the fact that we can
process transactions independently without messing with the
`purchaseCompleteCallbacksByProductID` state now.

### Future improvements
This is just a start, there's a lot more than can be done to simplify
the `PurchasesOrchestrator` monster. The main thing would be to remove
all the custom code for `restorePurchases`, which is largely the same as
what's in `TransactionPoster` now.

### Notes:
Just posting this here for posteriority. I had to trace the call
hierarchy for both SK1 and SK2. This "block" is what's now in
`TransactionPoster`:

![IMG_C60274393F7F-1](https://github.com/RevenueCat/purchases-ios/assets/685609/5a7377f0-8301-474f-a774-a90b6e232bc3)